Refactor Request Logistic Status
